+++ src/main/java/com/epam/esm/model/GiftCertificate.java
@@ -1,6 +1,5 @@
 package com.epam.esm.model;
 
-import org.hibernate.annotations.Where;
 import org.hibernate.envers.Audited;
 
 import javax.persistence.*;
@@ -19,7 +18,7 @@ public class GiftCertificate extends BaseModel {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @NotNull(message = "certificate must have name")
     @NotEmpty(message = "name must be not empty")
@@ -56,11 +55,11 @@ public class GiftCertificate extends BaseModel {
     )
     private Set<Tag> tags = new HashSet<>();
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
